chore: GuestbookResponseDto의 rating을 실수로 변경

+++ src/main/java/com/babzip/backend/guestbook/dto/response/GuestbookResponseDto.java
@@ -7,7 +7,7 @@ public record GuestbookResponseDto(
         String restaurantName,
         String kakaoPlaceId,
         String content,
-        Long rating,
+        Double rating,
         String address,
         LocalDateTime createdAt
 ) {
